myria: replace SingleGroupByAggregate with SingleGroupByAggregateNoBuffer
See discussion on #340
