+++ speedtest/edu/washington/escience/myria/mrbenchmarks/AggregateQueryVariantSQLite.java
@@ -13,7 +13,7 @@ import edu.washington.escience.myria.operator.DbQueryScan;
 import edu.washington.escience.myria.operator.RootOperator;
 import edu.washington.escience.myria.operator.SinkRoot;
 import edu.washington.escience.myria.operator.agg.Aggregator;
-import edu.washington.escience.myria.operator.agg.SingleGroupByAggregateNoBuffer;
+import edu.washington.escience.myria.operator.agg.SingleGroupByAggregate;
 import edu.washington.escience.myria.parallel.CollectConsumer;
 import edu.washington.escience.myria.parallel.CollectProducer;
 import edu.washington.escience.myria.parallel.ExchangePairID;
@@ -51,8 +51,8 @@ public class AggregateQueryVariantSQLite implements QueryPlanGenerator {
     final GenericShuffleConsumer sc =
         new GenericShuffleConsumer(shuffleLocalGroupBy.getSchema(), shuffleLocalGroupByID, allWorkers);
 
-    final SingleGroupByAggregateNoBuffer agg =
-        new SingleGroupByAggregateNoBuffer(sc, new int[] { 1 }, 0, new int[] { Aggregator.AGG_OP_SUM });
+    final SingleGroupByAggregate agg =
+        new SingleGroupByAggregate(sc, new int[] { 1 }, 0, new int[] { Aggregator.AGG_OP_SUM });
 
     final CollectProducer sendToMaster = new CollectProducer(agg, sendToMasterID, 0);
