+++ src/edu/washington/escience/myria/operator/agg/SingleGroupByAggregate.java
@@ -3,9 +3,8 @@ package edu.washington.escience.myria.operator.agg;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 
-import org.apache.commons.lang3.tuple.Pair;
+import org.joda.time.DateTime;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -17,6 +16,18 @@ import edu.washington.escience.myria.TupleBatchBuffer;
 import edu.washington.escience.myria.Type;
 import edu.washington.escience.myria.operator.Operator;
 import edu.washington.escience.myria.operator.UnaryOperator;
+import gnu.trove.iterator.TDoubleObjectIterator;
+import gnu.trove.iterator.TFloatObjectIterator;
+import gnu.trove.iterator.TIntObjectIterator;
+import gnu.trove.iterator.TLongObjectIterator;
+import gnu.trove.map.TDoubleObjectMap;
+import gnu.trove.map.TFloatObjectMap;
+import gnu.trove.map.TIntObjectMap;
+import gnu.trove.map.TLongObjectMap;
+import gnu.trove.map.hash.TDoubleObjectHashMap;
+import gnu.trove.map.hash.TFloatObjectHashMap;
+import gnu.trove.map.hash.TIntObjectHashMap;
+import gnu.trove.map.hash.TLongObjectHashMap;
 
 /**
  * The Aggregation operator that computes an aggregate (e.g., sum, avg, max, min) with a single group by column.
@@ -24,43 +35,84 @@ import edu.washington.escience.myria.operator.UnaryOperator;
 public class SingleGroupByAggregate extends UnaryOperator {
 
   /**
+   * The Logger.
+   */
+  private static final org.slf4j.Logger LOGGER = org.slf4j.LoggerFactory
+      .getLogger(SingleGroupByAggregate.class);
+
+  /**
    * default serialization ID.
-   * */
+   */
   private static final long serialVersionUID = 1L;
 
   /**
    * compute multiple aggregates in the same time. The columns to compute the aggregates are
-   * {@link SingleGroupByAggregate#afields}.
-   * */
-  private final Aggregator<?>[] agg;
+   * {@link SingleGroupByAggregateNoIntermediateBuffer#afields}.
+   */
+  private Aggregator<?>[] agg;
 
   /**
-   * Compute aggregate on each of the {@link SingleGroupByAggregate#afields}, with the corresponding {@link Aggregator}
-   * in @link SingleGroupByAggregate#agg}.
-   * */
+   * Compute aggregate on each of the {@link SingleGroupByAggregateNoIntermediateBuffer#afields}, with the corresponding
+   * {@link Aggregator} in @link SingleGroupByAggregate#agg}.
+   */
   private final int[] afields;
 
-  /** Aggregate operators. */
+  /**
+   * The aggregate operations that will be computed.
+   */
   private final int[] aggOps;
 
   /**
    * The group by column.
-   * */
+   */
   private final int gColumn;
 
   /**
-   * The buffer storing in-progress group by results. {groupby-column-value -> Aggregator Array}
-   * */
-  private transient HashMap<Object, Aggregator<?>[]> groupAggs;
+   * The group-by column type.
+   */
+  private Type gColumnType;
 
   /**
-   * {groupby-column-value -> }.
-   * */
-  private transient Map<Object, Pair<Object, TupleBatchBuffer>> groupedTupleBatches;
+   * aggregate column types.
+   */
+  private Type[] aColumnTypes;
 
   /**
-   * storing results after group by is done.
-   * */
+   * The buffer storing in-progress group by results. {groupby-column-value -> Aggregator Array} when the group key is
+   * String
+   */
+  private transient HashMap<String, Aggregator<?>[]> groupAggsString;
+
+  /**
+   * The buffer storing in-progress group by results. {groupby-column-value -> Aggregator Array} when the group key is
+   * DateTime.
+   */
+  private transient HashMap<DateTime, Aggregator<?>[]> groupAggsDatetime;
+
+  /**
+   * The buffer storing in-progress group by results when the group key is int.
+   */
+  private transient TIntObjectMap<Aggregator<?>[]> groupAggsInt;
+  /**
+   * The buffer storing in-progress group by results when the group key is boolean.
+   */
+  private transient Aggregator<?>[][] groupAggsBoolean;
+  /**
+   * The buffer storing in-progress group by results when the group key is long.
+   */
+  private transient TLongObjectMap<Aggregator<?>[]> groupAggsLong;
+  /**
+   * The buffer storing in-progress group by results when the group key is float.
+   */
+  private transient TFloatObjectMap<Aggregator<?>[]> groupAggsFloat;
+  /**
+   * The buffer storing in-progress group by results when the group key is double.
+   */
+  private transient TDoubleObjectMap<Aggregator<?>[]> groupAggsDouble;
+
+  /**
+   * The buffer storing results after group by is done.
+   */
   private transient TupleBatchBuffer resultBuffer;
 
   /**
@@ -77,105 +129,308 @@ public class SingleGroupByAggregate extends UnaryOperator {
     if (afields.length == 0) {
       throw new IllegalArgumentException("aggregation fields must not be empty");
     }
-
-    this.afields = afields;
     gColumn = gfield;
+    this.afields = afields;
     this.aggOps = aggOps;
-    agg = new Aggregator<?>[aggOps.length];
-    groupAggs = new HashMap<Object, Aggregator<?>[]>();
   }
 
   /**
    * @return the aggregate field
-   * */
+   */
   public final int[] aggregateFields() {
     return afields;
   }
 
   @Override
   protected final void cleanup() throws DbException {
-    groupAggs = new HashMap<Object, Aggregator<?>[]>();
-    groupedTupleBatches = new HashMap<Object, Pair<Object, TupleBatchBuffer>>();
+
+    groupAggsString = null;
+    groupAggsDouble = null;
+    groupAggsFloat = null;
+    groupAggsInt = null;
+    groupAggsLong = null;
     resultBuffer = null;
   }
 
   /**
-   * @param tb the TupleBatch to be processed.
-   * */
-  private void processTupleBatch(final TupleBatch tb) {
-    final Set<Pair<Object, TupleBatchBuffer>> readyTBB = tb.groupby(gColumn, groupedTupleBatches);
-    if (readyTBB != null) {
-      for (final Pair<Object, TupleBatchBuffer> p : readyTBB) {
-        final Object groupColumnValue = p.getKey();
-        final TupleBatchBuffer tbb = p.getRight();
-        Aggregator<?>[] groupAgg = groupAggs.get(groupColumnValue);
-        if (groupAgg == null) {
-          groupAgg = new Aggregator<?>[agg.length];
-          for (int j = 0; j < agg.length; j++) {
-            groupAgg[j] = agg[j].freshCopyYourself();
-          }
-          groupAggs.put(groupColumnValue, groupAgg);
-        }
+   * add a value to group by.
+   * 
+   * @param tb source TB.
+   * @param row which row.
+   * @param aggFieldIdx which column.
+   * @param aggFieldType the type of the value
+   * @param agg the aggregator.
+   */
+  public static void addValue2Group(final TupleBatch tb, final int row, final int aggFieldIdx, final Type aggFieldType,
+      final Aggregator<?> agg) {
+    switch (aggFieldType) {
+      case BOOLEAN_TYPE:
+        ((BooleanAggregator) agg).add(tb.getBoolean(aggFieldIdx, row));
+        break;
+      case INT_TYPE:
+        ((IntegerAggregator) agg).add(tb.getInt(aggFieldIdx, row));
+        break;
+      case LONG_TYPE:
+        ((LongAggregator) agg).add(tb.getLong(aggFieldIdx, row));
+        break;
+      case FLOAT_TYPE:
+        ((FloatAggregator) agg).add(tb.getFloat(aggFieldIdx, row));
+        break;
+      case DOUBLE_TYPE:
+        ((DoubleAggregator) agg).add(tb.getDouble(aggFieldIdx, row));
+        break;
+      case STRING_TYPE:
+        ((StringAggregator) agg).add(tb.getString(aggFieldIdx, row));
+        break;
+      case DATETIME_TYPE:
+        ((DateTimeAggregator) agg).add(tb.getDateTime(aggFieldIdx, row));
+        break;
 
-        TupleBatch filledTB = null;
-        while ((filledTB = tbb.popFilled()) != null) {
-          for (final Aggregator<?> ag : groupAgg) {
-            ag.add(filledTB);
-          }
-        }
-      }
     }
   }
 
   /**
-   * Finish processing, drain the buffers.
-   * */
-  private void processEnd() {
-    for (final Pair<Object, TupleBatchBuffer> p : groupedTupleBatches.values()) {
-      final Object groupColumnValue = p.getKey();
-      Aggregator<?>[] groupAgg = groupAggs.get(groupColumnValue);
-      if (groupAgg == null) {
-        groupAgg = new Aggregator<?>[agg.length];
-        for (int j = 0; j < agg.length; j++) {
-          groupAgg[j] = agg[j].freshCopyYourself();
+   * @param tb the TupleBatch to be processed.
+   */
+  private void processTupleBatch(final TupleBatch tb) {
+    switch (gColumnType) {
+      case BOOLEAN_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          boolean groupByKey = tb.getBoolean(gColumn, i);
+          Aggregator<?>[] groupAgg = null;
+          if (groupByKey) {
+            groupAgg = groupAggsBoolean[0];
+          } else {
+            groupAgg = groupAggsBoolean[1];
+          }
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            if (groupByKey) {
+              groupAggsBoolean[0] = groupAgg;
+            } else {
+              groupAggsBoolean[1] = groupAgg;
+            }
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
         }
-        groupAggs.put(groupColumnValue, groupAgg);
-      }
-      final TupleBatchBuffer tbb = p.getRight();
-      TupleBatch anyTBB = null;
-      while ((anyTBB = tbb.popAny()) != null) {
-        for (final Aggregator<?> ag : groupAgg) {
-          ag.add(anyTBB);
+        break;
+      case STRING_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          String groupByKey = tb.getString(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsString.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsString.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
         }
-      }
+        break;
+      case DATETIME_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          DateTime groupByKey = tb.getDateTime(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsDatetime.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsDatetime.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
+        }
+        break;
+      case INT_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          int groupByKey = tb.getInt(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsInt.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsInt.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
+        }
+        break;
+      case LONG_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          long groupByKey = tb.getLong(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsLong.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsLong.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
+        }
+        break;
+      case FLOAT_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          float groupByKey = tb.getFloat(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsFloat.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsFloat.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
+        }
+        break;
+      case DOUBLE_TYPE:
+        for (int i = 0; i < tb.numTuples(); i++) {
+          double groupByKey = tb.getDouble(gColumn, i);
+          Aggregator<?>[] groupAgg = groupAggsDouble.get(groupByKey);
+          if (groupAgg == null) {
+            groupAgg = new Aggregator<?>[agg.length];
+            for (int j = 0; j < agg.length; j++) {
+              groupAgg[j] = agg[j].freshCopyYourself();
+            }
+            groupAggsDouble.put(groupByKey, groupAgg);
+          }
+          for (int j = 0; j < afields.length; j++) {
+            addValue2Group(tb, i, afields[j], aColumnTypes[j], groupAgg[j]);
+          }
+        }
+        break;
     }
+
   }
 
   /**
    * @param resultBuffer where the results are stored.
-   * */
+   */
   private void generateResult(final TupleBatchBuffer resultBuffer) {
 
-    for (final Map.Entry<Object, Aggregator<?>[]> e : groupAggs.entrySet()) {
-      final Object groupByValue = e.getKey();
-      final Aggregator<?>[] aggLocal = e.getValue();
-      resultBuffer.put(0, groupByValue);
-      int fromIndex = 1;
-      for (final Aggregator<?> element : aggLocal) {
-        element.getResult(resultBuffer, fromIndex);
-        fromIndex += element.getResultSchema().numColumns();
-      }
+    switch (gColumnType) {
+      case BOOLEAN_TYPE:
+        Aggregator<?>[] t = groupAggsBoolean[0];
+        if (t != null) {
+          resultBuffer.putBoolean(0, true);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : t) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        Aggregator<?>[] f = groupAggsBoolean[1];
+        if (f != null) {
+          resultBuffer.putBoolean(1, true);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : f) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
+      case STRING_TYPE:
+        for (final Map.Entry<String, Aggregator<?>[]> e : groupAggsString.entrySet()) {
+          final String groupByValue = e.getKey();
+          final Aggregator<?>[] aggLocal = e.getValue();
+          resultBuffer.putString(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
+      case DATETIME_TYPE:
+        for (final Map.Entry<DateTime, Aggregator<?>[]> e : groupAggsDatetime.entrySet()) {
+          final DateTime groupByValue = e.getKey();
+          final Aggregator<?>[] aggLocal = e.getValue();
+          resultBuffer.putDateTime(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
+      case INT_TYPE:
+        TIntObjectIterator<Aggregator<?>[]> itInt = groupAggsInt.iterator();
+        while (itInt.hasNext()) {
+          itInt.advance();
+          int groupByValue = itInt.key();
+          final Aggregator<?>[] aggLocal = itInt.value();
+          resultBuffer.putInt(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
+      case LONG_TYPE:
+        TLongObjectIterator<Aggregator<?>[]> itLong = groupAggsLong.iterator();
+        while (itLong.hasNext()) {
+          itLong.advance();
+          long groupByValue = itLong.key();
+          final Aggregator<?>[] aggLocal = itLong.value();
+          resultBuffer.putLong(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        groupAggsLong = new TLongObjectHashMap<Aggregator<?>[]>();
+        break;
+      case FLOAT_TYPE:
+        groupAggsFloat = new TFloatObjectHashMap<Aggregator<?>[]>();
+        TFloatObjectIterator<Aggregator<?>[]> itFloat = groupAggsFloat.iterator();
+        while (itFloat.hasNext()) {
+          itFloat.advance();
+          float groupByValue = itFloat.key();
+          final Aggregator<?>[] aggLocal = itFloat.value();
+          resultBuffer.putFloat(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
+      case DOUBLE_TYPE:
+        TDoubleObjectIterator<Aggregator<?>[]> itDouble = groupAggsDouble.iterator();
+        while (itDouble.hasNext()) {
+          itDouble.advance();
+          double groupByValue = itDouble.key();
+          final Aggregator<?>[] aggLocal = itDouble.value();
+          resultBuffer.putDouble(0, groupByValue);
+          int fromIndex = 1;
+          for (final Aggregator<?> element : aggLocal) {
+            element.getResult(resultBuffer, fromIndex);
+            fromIndex += element.getResultSchema().numColumns();
+          }
+        }
+        break;
     }
+
   }
 
-  /**
-   * Returns the next tuple. If there is a group by field, then the first field is the field by which we are grouping,
-   * and the second field is the result of computing the aggregate, If there is no group by field, then the result tuple
-   * should contain one field representing the result of the aggregate. Should return null if there are no more tuples.
-   * 
-   * @throws DbException if processing error occurs.
-   * @return next TB.
-   */
   @Override
   protected final TupleBatch fetchNextReady() throws DbException {
     TupleBatch tb = null;
@@ -190,66 +445,78 @@ public class SingleGroupByAggregate extends UnaryOperator {
     }
 
     while ((tb = child.nextReady()) != null) {
+
+      if (LOGGER.isDebugEnabled()) {
+        LOGGER.debug("get a TB from child");
+      }
       processTupleBatch(tb);
     }
 
     if (child.eos() || child.eoi()) {
-      processEnd();
       generateResult(resultBuffer);
     }
     return resultBuffer.popAny();
   }
 
-  @Override
-  protected final Schema generateSchema() {
-    return generateSchema(getChild(), groupAggs, gColumn, afields, agg, aggOps);
-  }
-
   /**
    * @return the group by column.
-   * */
+   */
   public final int getGroupByColumn() {
     return gColumn;
   }
 
   @Override
   protected final void init(final ImmutableMap<String, Object> execEnvVars) throws DbException {
-    groupAggs = new HashMap<Object, Aggregator<?>[]>();
-    groupedTupleBatches = new HashMap<Object, Pair<Object, TupleBatchBuffer>>();
+    generateSchema();
+    switch (gColumnType) {
+      case BOOLEAN_TYPE:
+        groupAggsBoolean = new Aggregator<?>[2][];
+        break;
+      case INT_TYPE:
+        groupAggsInt = new TIntObjectHashMap<Aggregator<?>[]>();
+        break;
+      case LONG_TYPE:
+        groupAggsLong = new TLongObjectHashMap<Aggregator<?>[]>();
+        break;
+      case FLOAT_TYPE:
+        groupAggsFloat = new TFloatObjectHashMap<Aggregator<?>[]>();
+        break;
+      case DOUBLE_TYPE:
+        groupAggsDouble = new TDoubleObjectHashMap<Aggregator<?>[]>();
+        break;
+      case STRING_TYPE:
+        groupAggsString = new HashMap<String, Aggregator<?>[]>();
+        break;
+      case DATETIME_TYPE:
+        groupAggsDatetime = new HashMap<DateTime, Aggregator<?>[]>();
+        break;
+    }
     resultBuffer = new TupleBatchBuffer(getSchema());
+
   }
 
-  /**
-   * Generates the schema for SingleGroupByAggregate.
-   * 
-   * @param child the child operator
-   * @param groupAggs the aggregator for each grouping
-   * @param gfield the group field
-   * @param afields the aggregate fields
-   * @param agg the aggregators to be populated
-   * @param aggOps ints representing the aggregate operations
-   * 
-   * @return the schema based on the aggregate, if the child is null, will return null.
-   */
-  private static Schema generateSchema(final Operator child, final Map<Object, Aggregator<?>[]> groupAggs,
-      final int gfield, final int[] afields, final Aggregator<?>[] agg, final int[] aggOps) {
-    if (child == null) {
-      return null;
+  @Override
+  protected Schema generateSchema() {
+    final Schema childSchema = getChild().getSchema();
+    if (gColumn < 0 || gColumn >= childSchema.numColumns()) {
+      throw new IllegalArgumentException("Invalid group field");
     }
+
     Schema outputSchema = null;
 
-    final Schema childSchema = child.getSchema();
-    final ImmutableList.Builder<Type> gTypes = ImmutableList.builder();
-    gTypes.add(childSchema.getColumnType(gfield));
-    final ImmutableList.Builder<String> gNames = ImmutableList.builder();
-    gNames.add(childSchema.getColumnName(gfield));
+    outputSchema =
+        new Schema(ImmutableList.of(childSchema.getColumnType(gColumn)), ImmutableList.of(childSchema
+            .getColumnName(gColumn)));
 
-    // Generates the output schema
-    outputSchema = new Schema(gTypes, gNames);
+    gColumnType = childSchema.getColumnType(gColumn);
+    aColumnTypes = new Type[afields.length];
+
+    agg = new Aggregator<?>[aggOps.length];
 
     int idx = 0;
     for (final int afield : afields) {
-      switch (childSchema.getColumnType(afield)) {
+      aColumnTypes[idx] = childSchema.getColumnType(afield);
+      switch (aColumnTypes[idx]) {
         case BOOLEAN_TYPE:
           agg[idx] = new BooleanAggregator(afield, childSchema.getColumnName(afield), aggOps[idx]);
           break;
@@ -268,12 +535,10 @@ public class SingleGroupByAggregate extends UnaryOperator {
         case STRING_TYPE:
           agg[idx] = new StringAggregator(afield, childSchema.getColumnName(afield), aggOps[idx]);
           break;
-        case DATETIME_TYPE:
-          agg[idx] = new DateTimeAggregator(afield, childSchema.getColumnName(afield), aggOps[idx]);
-          break;
         default:
-          throw new IllegalArgumentException("Unknown column type: " + childSchema.getColumnType(afield));
+          throw new IllegalArgumentException("Unknown column type: " + aColumnTypes[idx]);
       }
+
       outputSchema = Schema.merge(outputSchema, agg[idx].getResultSchema());
       idx++;
     }
